Corrijindo testes após refatoração e otimização

+++ IntegradorTesteUnidade/AplicacaoTestes/PipelineAplicacaoTestes/ExecutorPipelineTestes/FeatureExecutorTestes/ManipulacaoDadosTestes/RemoverColunaExecutorTests.cs
@@ -1,9 +1,6 @@
 ﻿using IntegradorAplicacao.PipelineAplicacao.ExecutorPipeline.FeatureExecutor.ManipulacaoDados;
 using IntegradorDominio.DataFrameModel;
 using IntegradorDominio.FeatureEngineering.ManipulacaoDados;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.ExecutorPipelineTestes.FeatureExecutorTestes.ManipulacaoDadosTestes
 {
@@ -13,7 +10,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
         public void DeveRemoverUmaColuna()
         {
             var df = new DataFrame();
-            df.AdicionarColuna<int?>("Id", new List<int?> { 1, 2 });
+            df.AdicionarColuna<float?>("Id", new List<float?> { 1, 2 });
             df.AdicionarColuna<string?>("Nome", new List<string?> { "A", "B" });
 
             var operacao = new RemoverColuna { col = "[Id]" };
@@ -21,7 +18,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
 
             var resultado = (DataFrame)executor.Executar(df);
 
-            Assert.Null(resultado.PegarColuna<int?>("Id"));
+            Assert.Null(resultado.PegarColuna<float?>("Id"));
             Assert.NotNull(resultado.PegarColuna<string>("Nome"));
             Assert.Equal(2, resultado.QuantidadeLinhas);
         }
@@ -30,7 +27,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
         public void DeveRemoverMultiplasColunas()
         {
             var df = new DataFrame();
-            df.AdicionarColuna<int?>("Id", new List<int?> { 1 });
+            df.AdicionarColuna<float?>("Id", new List<float?> { 1 });
             df.AdicionarColuna<string?>("Nome", new List<string?> { "A" });
             df.AdicionarColuna<bool?>("Flag", new List<bool?> { true });
 
@@ -39,7 +36,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
 
             var resultado = (DataFrame)executor.Executar(df);
 
-            Assert.Null(resultado.PegarColuna<int?>("Id"));
+            Assert.Null(resultado.PegarColuna<float?>("Id"));
             Assert.Null(resultado.PegarColuna<bool?>("Flag"));
             Assert.NotNull(resultado.PegarColuna<string>("Nome"));
         }
@@ -48,7 +45,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
         public void NaoRemoveNadaSeColunaNaoExistir()
         {
             var df = new DataFrame();
-            df.AdicionarColuna<int?>("Id", new List<int?> { 1 });
+            df.AdicionarColuna<float?>("Id", new List<float?> { 1 });
             df.AdicionarColuna<string?>("Nome", new List<string?> { "A" });
 
             var operacao = new RemoverColuna { col = "[NaoExiste]" };
@@ -56,7 +53,7 @@ namespace IntegradorTesteUnidade.AplicacaoTestes.PipelineAplicacaoTestes.Executo
 
             var resultado = (DataFrame)executor.Executar(df);
 
-            Assert.NotNull(resultado.PegarColuna<int?>("Id"));
+            Assert.NotNull(resultado.PegarColuna<float?>("Id"));
             Assert.NotNull(resultado.PegarColuna<string>("Nome"));
             Assert.Equal(1, resultado.QuantidadeLinhas);
         }
